last update

+++ app/Http/Controllers/BiographyController.php
@@ -298,51 +298,54 @@ class BiographyController extends Controller
         $updated_data = [];
         // dd($gambarSertifikat);
 
-
-        if (count($gambarSertifikat) > 0) {
-
-            for ($i = 0; $i < count($gambarSertifikat); $i++) {
-
-                $sertifikat = $request->sertifikat[$i];
-                $foto_sertifikat = $gambarSertifikat[$i];
-
-                // Cek apakah data sebelumnya ada atau tidak
-                if (isset($biography->sertifikat[$i])) {
-
-                    // Cek apakah data diubah atau tidak
-                    if ($biography->sertifikat[$i]['gambar'] == $foto_sertifikat['gambar']) {
-                        // Sama
-                        array_push($updated_data, $foto_sertifikat);
+        if ($gambarSertifikat) {
+            
+            if (count($gambarSertifikat) > 0) {
+    
+                for ($i = 0; $i < count($gambarSertifikat); $i++) {
+    
+                    $sertifikat = $request->sertifikat[$i];
+                    $foto_sertifikat = $gambarSertifikat[$i];
+    
+                    // Cek apakah data sebelumnya ada atau tidak
+                    if (isset($biography->sertifikat[$i])) {
+    
+                        // Cek apakah data diubah atau tidak
+                        if ($biography->sertifikat[$i]['gambar'] == $foto_sertifikat['gambar']) {
+                            // Sama
+                            array_push($updated_data, $foto_sertifikat);
+                        } else {
+                            // Jika Berubah maka hapus file path lama
+                            File::delete(public_path($biography->sertifikat[$i]['gambar']));
+    
+                            // Upload file yang berubah
+                            $nama_foto_sertifikat = 'sertifikat_atlet/binawa_foto_sertifikat_' . $i . '_' . date('Ymdhis') . '.' . $foto_sertifikat['gambar']->getClientOriginalExtension();
+                            $foto_sertifikat['gambar']->move('sertifikat_atlet', $nama_foto_sertifikat);
+                            $sertifikat['gambar'] = $nama_foto_sertifikat;
+                            array_push($updated_data, $sertifikat);
+                        }
+                        // dd("ada");
                     } else {
-                        // Jika Berubah maka hapus file path lama
-                        File::delete(public_path($biography->sertifikat[$i]['gambar']));
-
-                        // Upload file yang berubah
+                        // dd("Tidak Ada");
+                        // Tidak ada
+                        // Upload File Baru
                         $nama_foto_sertifikat = 'sertifikat_atlet/binawa_foto_sertifikat_' . $i . '_' . date('Ymdhis') . '.' . $foto_sertifikat['gambar']->getClientOriginalExtension();
                         $foto_sertifikat['gambar']->move('sertifikat_atlet', $nama_foto_sertifikat);
                         $sertifikat['gambar'] = $nama_foto_sertifikat;
                         array_push($updated_data, $sertifikat);
                     }
-                    // dd("ada");
-                } else {
-                    // dd("Tidak Ada");
-                    // Tidak ada
-                    // Upload File Baru
-                    $nama_foto_sertifikat = 'sertifikat_atlet/binawa_foto_sertifikat_' . $i . '_' . date('Ymdhis') . '.' . $foto_sertifikat['gambar']->getClientOriginalExtension();
-                    $foto_sertifikat['gambar']->move('sertifikat_atlet', $nama_foto_sertifikat);
-                    $sertifikat['gambar'] = $nama_foto_sertifikat;
-                    array_push($updated_data, $sertifikat);
+                }
+                $validatedData['sertifikat'] = $updated_data;
+            } else {
+                // Kosong
+                // dd("Kosong");
+                foreach ($biography->sertifikat as $sertifikat) {
+                    // Hapus file
+                    File::delete(public_path($sertifikat['gambar']));
                 }
             }
-            $validatedData['sertifikat'] = $updated_data;
-        } else {
-            // Kosong
-            // dd("Kosong");
-            foreach ($biography->sertifikat as $sertifikat) {
-                // Hapus file
-                File::delete(public_path($sertifikat['gambar']));
-            }
         }
+
 
 
         $validatedData2 = array_replace($validatedData, $replacement);
